requirements.txt
azure-core==1.25.1
azure-data-tables==12.4.0
azure-functions==1.11.2
azure-storage-blob==12.13.1
Flask==2.2.2
Flask-RESTful==0.3.9
mypy-extensions==0.4.3
numpy==1.22.4
openscad-runner==1.1.0
pipenv==2022.9.8
pipreqs==0.4.11
py==1.11.0
PyYAML==6.0
shortuuid==1.0.9
